Tech-debt cleanups: useHiddenList, buildXAxisConfig, HgChildPane (#1089)
Extract a shared useHiddenList hook (useRowManagement becomes a thin wrapper;
useColumnManagement keeps its sort coupling), pull XYChart's X-axis config into
a pure, tested xychart-axis.ts, and move SortableChild/HgChildPane into a
sibling file. Skipped the materializeTable and query-cell-metadata factories
(premises did not hold on inspection) and verified the macro-substitution regex
factories are already DRY; rationale recorded in the plan.
